Add FetchError class to utils/errors.ts
A new class FetchError has been added to handle errors during fetching. It extends the Error class and takes in an Error object, a Request object, and an optional Response object as arguments. The aim is to provide detailed error instances including request and response data, hence improving the debugging process.